eval/requirements.txt
azure-ai-evaluation>=1.0.0
python-dotenv>=1.0.0
